Fix Handler call
