+++ Covid19Radar/Covid19Radar/Covid19Radar.csproj
@@ -10,6 +10,8 @@
     <TargetFramework>netstandard2.1</TargetFramework>
     <LangVersion>8.0</LangVersion>
     <NeutralLanguage>en-US</NeutralLanguage>
+    <Copyright />
+    <PackageProjectUrl></PackageProjectUrl>
   </PropertyGroup>
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets" Label="MultilingualAppToolkit" Condition="Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\v$(MultilingualAppToolkitVersion)\Microsoft.Multilingual.ResxResources.targets')" />
   <Target Name="MATPrerequisite" BeforeTargets="PrepareForBuild" Condition="!Exists('$(MSBuildExtensionsPath)\Microsoft\Multilingual App Toolkit\Microsoft.Multilingual.ResxResources.targets')" Label="MultilingualAppToolkit">
@@ -57,6 +59,7 @@
     <None Remove="Resources\Fonts\NotoSansCJKjp-Medium.otf" />
     <None Remove="Resources\Fonts\NotoSansCJKjp-Regular.otf" />
     <None Remove="Resources\Fonts\Roboto-Regular.ttf" />
+    <None Remove="Services\ExposureNotificationHandler.cs.txt" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\Fonts\fa-brands-400.otf" />
