+++ Covid19Radar/Covid19Radar/Services/HttpDataService.cs
@@ -102,7 +102,6 @@ namespace Covid19Radar.Services
             return null;
         }
 
-        /* TOOD Marge EN
         public async Task<TemporaryExposureKeysHttpResultModel> GetTemporaryExposureKeys(long since)
         {
             string url = AppConstants.ApiBaseUrl + $"/TemporaryExposureKeys?since={since}";
@@ -113,7 +112,6 @@ namespace Covid19Radar.Services
             }
             return null;
         }
-        */
 
         // GET /api/notification/pull/{lastClientUpdateTime:datetime} - pull Notifications 
         public async Task<NotificationPullResult> GetNotificationPullAsync(UserDataModel user)
