+++ Covid19Radar/Covid19Radar/Services/TestNativeImplementation.cs
@@ -1,5 +1,4 @@
-﻿#if DEBUG
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
@@ -74,4 +73,3 @@ namespace Covid19Radar.Services
 		}
 	}
 }
-#endif
